requirements.txt
streamlit
manim
python-dotenv
google-generativeai
manim-voiceover[gtts]
moviepy
openai
